chore(android,mobile): rename Android package to ai.elizaos.app; on-device bundle fixes
Pre-existing working-tree changes committed alongside the e2e work so they
are not lost.

- Rename the Android app package ai.milady.milady -> ai.elizaos.app across all
  Java sources, AndroidManifest, build.gradle, string/shortcut resources, and
  the electrobun preload (content-preserving moves).
- build-mobile-bundle: force @elizaos/agent and @elizaos/app-core to resolve
  from src so their circular barrel re-exports inline correctly under Bun
  (dist re-exports came through undefined on-device).
- agent api/server: guard handleCloudPairRoute — it can be undefined in the
  on-device bundle (circular re-export init order) and an unguarded call
  500s every request, wedging the dashboard at "Connecting to backend…".
- drop stale plugin .build-lock/metadata.json artifacts.
