+++ packages/app-core/platforms/android/app/src/main/java/ai/elizaos/app/ElizaAssistActivity.java
@@ -1,4 +1,4 @@
-package ai.milady.milady;
+package ai.elizaos.app;
 
 import android.app.Activity;
 import android.content.Intent;
