+++ packages/app-core/platforms/android/app/src/main/java/ai/elizaos/app/ElizaBootReceiver.java
@@ -1,4 +1,4 @@
-package ai.milady.milady;
+package ai.elizaos.app;
 
 import android.app.AppOpsManager;
 import android.content.BroadcastReceiver;
