+++ packages/app-core/platforms/android/app/src/main/java/ai/elizaos/app/ElizaInCallService.java
@@ -1,4 +1,4 @@
-package ai.milady.milady;
+package ai.elizaos.app;
 
 import android.content.Intent;
 import android.net.Uri;
