+++ packages/app-core/platforms/android/app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <?xml version='1.0' encoding='utf-8'?>
 <resources>
-    <string name="app_name">Milady</string>
-    <string name="title_activity_main">Milady</string>
-    <string name="package_name">ai.milady.milady</string>
-    <string name="custom_url_scheme">milady</string>
+    <string name="app_name">Eliza</string>
+    <string name="title_activity_main">Eliza</string>
+    <string name="package_name">ai.elizaos.app</string>
+    <string name="custom_url_scheme">elizaos</string>
 </resources>
