--- plugins/plugin-native-system/.build-lock/metadata.json
@@ -1,9 +1,0 @@
-{
-  "pid": 561,
-  "command": [
-    "bun",
-    "run",
-    "build:unlocked"
-  ],
-  "createdAt": "2026-06-02T15:11:59.546Z"
-}
