--- plugins/plugin-native-wifi/.build-lock/metadata.json
@@ -1,9 +1,0 @@
-{
-  "pid": 309,
-  "command": [
-    "bun",
-    "run",
-    "build:unlocked"
-  ],
-  "createdAt": "2026-06-02T15:11:59.330Z"
-}
